refactor(streaming): clean up service option generation
